+++ src/features/auth/ui/signIn/signIn.tsx
@@ -1,7 +1,7 @@
 import { Controller, FieldErrors } from 'react-hook-form'
 
 import { AuthCard } from '@/features'
-import { ROUTES } from '@/shared/config/routes'
+import { ROUTES } from '@/shared/config'
 import { Button, Input, Typography } from '@photo-fiesta/ui-lib'
 import Link from 'next/link'
