+++ src/features/auth/ui/signUp/useSignUpForm.tsx
@@ -2,8 +2,8 @@ import { useState } from 'react'
 import { useForm } from 'react-hook-form'
 import { toast } from 'react-toastify'
 
-import { useSignUpMutation } from '@/features/auth'
-import { PASSWORD_REGEX, USERNAME_REGEX } from '@/shared/config/regex-constants'
+import { useSignUpMutation } from '@/features'
+import { PASSWORD_REGEX, USERNAME_REGEX } from '@/shared/config'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { z } from 'zod'
