+++ src/shared/config/index.ts
@@ -1,2 +1,3 @@
+export * from './apiURLs'
 export * from './regex-constants'
 export * from './routes'
